requirements.txt
branca==0.4.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
folium==0.11.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.2
python-dotenv==0.14.0
requests==2.24.0
rethinkdb==2.4.7
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
